requirements/base.txt
Django~=2.2.9
django-filter==2.2.0
django-health-check==3.8.0
git+https://github.com/Humanitec/django-oauth-toolkit-jwt@v0.5.2#egg=django-oauth-toolkit-jwt
django-oauth-toolkit==1.3.2
djangorestframework==3.9.4
psycopg2-binary==2.8.3
drf-yasg==1.10.2
voluptuous
factory_boy==3.0.1
Pillow==2.9.0
idna==2.6
requests==2.25.1
